+++ .gitignore
@@ -60,3 +60,6 @@ lib/**/.gitignore
 # cache dirs
 tools/fdcanusb
 
+# auto-generated config headers
+**/*_config.hpp
+
