+++ lib/CMakeLists.txt
@@ -10,7 +10,7 @@ set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-add_subdirectory(config_gen)
+add_subdirectory(config)
 add_subdirectory(dbc)
 if (DEFINED MX_SRC_DIR AND DEFINED MX_INC_DIR AND DEFINED MX_STARTUP_S)
     add_subdirectory(stm32g4)
