+++ lib/config/templates/config_header.hpp.j2
@@ -1,0 +1,144 @@
+#pragma once
+
+// ======================================================
+// THIS FILE IS AUTO-GENERATED BY config_gen.py
+// Last Updated: {{ timestamp }}
+// ======================================================
+
+#include <serial/fdcan.hpp>
+#include <serial/uart.hpp>
+#include <tuple>
+
+#include <CANBus1.hpp>
+#include <adc.hpp>
+#include <hw/ad8418a.hpp>
+#include <hw/flash.hpp>
+
+namespace mrover {
+
+    struct {{project_name}}_config_t {
+
+        static inline void* flash_ptr = nullptr;
+
+        {% set num_filters = (can_filtering.can_subs | default([]) | length) + 1 %}
+        {% if num_filters != 0 %}
+        FDCAN::Filter can_node_filters[{{num_filters}}];
+        {% endif %}
+
+{{ reg_output }}
+{{ field_output }}
+
+
+        template<typename F>
+        auto get() const { return F::get(*this); }
+
+        template<typename F>
+        void set(auto value) { F::set(*this, value); }
+
+        constexpr auto all() {
+            return std::forward_as_tuple(
+                {% for reg in reg_names %}
+                {{ reg }}{% if not loop.last %},{% endif %}
+
+                {% endfor %}            );
+        }
+
+        constexpr auto all() const {
+            return std::forward_as_tuple(
+                {% for reg in reg_names %}
+                {{ reg }}{% if not loop.last %},{% endif %}
+
+                {% endfor %}            );
+        }
+
+        auto set_raw(uint8_t address, uint32_t const raw) -> bool {
+            bool found = false;
+
+            std::apply(
+                    [&](auto const&... reg) -> void {
+                        (
+                                [&] -> void {
+                                    if (reg.addr == address) {
+                                        using T = std::remove_reference_t<decltype(reg)>::value_t;
+                                        reg.write(*this, from_raw<T>(raw));
+                                        found = true;
+                                    }
+                                }(),
+                                ...);
+                    },
+                    all());
+
+            return found;
+        }
+
+        auto get_raw(uint8_t address, uint32_t& raw) const -> bool {
+            bool found = false;
+            std::apply([&](auto const&... reg) -> void {
+                ((reg.addr == address ? (raw = to_raw(reg.value.value_or(0)), found = true) : false), ...);
+            },
+                        all());
+            return found;
+        }
+
+        struct mem_layout {
+            static constexpr uint32_t FLASH_BEGIN_ADDR = {{ flash_begin }};
+            static constexpr uint32_t FLASH_END_ADDR = {{ flash_end }};
+            static constexpr int PAGE_SIZE = {{flash_page_size}};
+            static constexpr int NUM_PAGES = {{flash_num_pages}};
+        };
+
+        static consteval auto size_bytes() -> uint16_t {
+            return validated_config_t<{{ project_name }}_config_t>::size_bytes();
+        }
+
+    }; // {{project_name}}_config_t
+
+    /**
+     * Get the {{ project_name }} CAN settings.
+     *
+     * Delay compensation needs to be enabled to allow BRS.
+     *
+     * @return CAN options for {{ project_name }}
+     */
+    inline auto get_can_options({{ project_name }}_config_t* config) -> FDCAN::Options {
+
+        {% if can_filtering is none %}
+
+        auto can_opts = FDCAN::Options{};
+        can_opts.delay_compensations = true;
+        can_opts.tdc_offset = 13;
+        can_opts.tdc_filter = 1;
+        return can_opts;
+
+        {% else %}
+        config->can_node_filters[0].id1 = config->get<{{ project_name }}_config_t::{{can_filtering.id_reg }}>();
+        config->can_node_filters[0].id2 = CAN_DEST_ID_MASK;
+        config->can_node_filters[0].id_type = FDCAN::FilterIdType::{{ can_filtering.id_type }};
+        config->can_node_filters[0].action = FDCAN::ActionType::Accept;
+        config->can_node_filters[0].mode = FDCAN::ActionType::Mask;
+        {% for sub in can_filtering.can_subs | default([]) %}
+
+        config->can_node_filters[{{ loop.index }}].id1 =  {{ sub.can_id }};
+        config->can_node_filters[{{ loop.index }}].id2 = CAN_SRC_ID_MASK;
+        config->can_node_filters[{{ loop.index }}].id_type = FDCAN::FilterIdType::{{ sub.id_type }};
+        config->can_node_filters[{{ loop.index }}].action = FDCAN::ActionType::Accept;
+        config->can_node_filters[{{ loop.index }}].mode = FDCAN::ActionType::Mask;
+        {% endfor %}
+
+        FDCAN::FilterConfig filter;
+        filter.begin = &config->can_node_filter;
+        filter.end = &config->can_node_filter + {{ (can_filtering.can_subs | default([]) | length) + 1}};
+        filter.global_non_matching_std_action = FDCAN::FilterAction::Reject;
+        filter.global_non_matching_ext_action = FDCAN::FilterAction::Reject;
+
+        auto can_opts = FDCAN::Options{};
+        can_opts.delay_compensation = {{ can_filtering.delay_compensation | lower }};
+        can_opts.tdc_offset = {{ can_filtering.tdc_offset }};
+        can_opts.tdc_filter = {{ can_filtering.tdc_filter }};
+        can_opts.filter_config = filter;
+        return can_opts;
+
+        {% endif %}
+    } // get_can_options
+
+} // namespace mrover
