--- lib/config_gen/templates/config_header.hpp.j2
@@ -1,91 +1,0 @@
-#pragma once
-
-// ======================================================
-// THIS FILE IS AUTO-GENERATED BY config_gen.py
-// Last Updated: {{ timestamp }}
-// ======================================================
-
-#include <serial/fdcan.hpp>
-#include <serial/uart.hpp>
-#include <tuple>
-
-#include <CANBus1.hpp>
-#include <adc.hpp>
-#include <hw/ad8418a.hpp>
-#include <hw/flash.hpp>
-
-namespace mrover {
-
-    struct {{struct_name}} {
-
-        static inline void* flash_ptr = nullptr;
-
-        {% if num_filters != 0 %}
-        FDCAN::Filter can_node_filter[{{num_filters}}];
-        {% endif %}
-
-{{ reg_output }}
-{{ field_output }}
-
-
-        template<typename F>
-        auto get() const { return F::get(*this); }
-
-        template<typename F>
-        void set(auto value) { F::set(*this, value); }
-
-
-{{ all_fn }}
-
-        auto set_raw(uint8_t address, uint32_t const raw) -> bool {
-            bool found = false;
-
-            std::apply(
-                    [&](auto const&... reg) -> void {
-                        (
-                                [&] -> void {
-                                    if (reg.addr == address) {
-                                        using T = std::remove_reference_t<decltype(reg)>::value_t;
-                                        reg.write(*this, from_raw<T>(raw));
-                                        found = true;
-                                    }
-                                }(),
-                                ...);
-                    },
-                    all());
-
-            return found;
-        }
-
-        auto get_raw(uint8_t address, uint32_t& raw) const -> bool {
-            bool found = false;
-            std::apply([&](auto const&... reg) -> void {
-                ((reg.addr == address ? (raw = to_raw(reg.value.value_or(0)), found = true) : false), ...);
-            },
-                        all());
-            return found;
-        }
-
-        struct mem_layout {
-            static constexpr uint32_t FLASH_BEGIN_ADDR = {{ flash_begin }};
-            static constexpr uint32_t FLASH_END_ADDR = {{ flash_end }};
-            static constexpr int PAGE_SIZE = {{flash_page_size}};
-            static constexpr int NUM_PAGES = {{flash_num_pages}};
-        };
-
-        static consteval auto size_bytes() -> uint16_t {
-            return validated_config_t<bmc_config_t>::size_bytes();
-        }
-
-    }; // {{struct_name}}
-
-    /**
-     * Get the BMC CAN settings.
-     *
-     * Delay compensation needs to be enabled to allow BRS.
-     *
-     * @return CAN options for BMC
-     */
-{{ can_fn }}
-
-} // namespace mrover
