+++ src/bmc/CMakeLists.txt
@@ -81,7 +81,7 @@ target_link_libraries(${CMAKE_PROJECT_NAME}
     stm32
     dbc
     util
-    config_gen
+    config
 )
 
 # Link compile commands for LSPs
